pyproject.toml
[tool.poetry]
name = "sysfacts"
version = "0.1.0"
description = ""
authors = ["Panos Mavrogiorgos <pmav99@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
py-cpuinfo = "^4.0"
psutil = "^5.4"
pendulum = "^2.0"
distro = "^1.3"
click = "^7.0"
"ruamel.yaml" = "^0.15.86"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
mypy = "^0.660.0"
black = {version = "^18.3-alpha.0",allows-prereleases = true}
pudb = "^2018.1"
ipython = "^7.2"

[tool.poetry.scripts]
sysfacts = "sysfacts.cli:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
